move moccml metaprog plugin to proper place
deploye moccmlxdsml.metaprog in the studio
