+++ ccsljava_xdsml/plugins/org.eclipse.gemoc.execution.moccmlxdsml.metaprog/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>org.eclipse.gemoc.addon.metaprogramming</name>
+	<name>org.eclipse.gemoc.execution.moccmlxdsml.metaprog</name>
 	<comment></comment>
 	<projects>
 	</projects>
