+++ ccsljava_xdsml/plugins/org.eclipse.gemoc.execution.moccmlxdsml.metaprog/plugin.xml
@@ -5,19 +5,18 @@
          point="org.eclipse.gemoc.gemoc_language_workbench.metaprog">
       <approach
             dependencies="org.eclipse.gemoc.metaprog.ecore"
-            name="org.eclipse.gemoc.metaprog.moccml"
-            validator="org.eclipse.gemoc.addon.metaprogramming.MOCCMLRuleProvider">
+            name="org.eclipse.gemoc.metaprog.moccml">
             <languageComponent
                   description="Entry used to define the ecl file in which the MOCCML elements are defined."
                   name="ecl"
                   optional="false"
-                  validationRule="org.eclipse.gemoc.addon.metaprogramming.MOCCMLRule">
+                  validationRule="org.eclipse.gemoc.execution.moccmlxdsml.metaprog.MOCCMLRule">
          </languageComponent>
             <languageComponent
                   description="Entry used to define the domain specific actions. Unlike for Kermeta3, no tag Main or InitializeModel are required."
                   name="k3"
                   optional="false"
-                  validationRule="org.eclipse.gemoc.addon.metaprogramming.Kermeta3Rule">
+                  validationRule="org.eclipse.gemoc.execution.moccmlxdsml.metaprog.Kermeta3Rule">
             </languageComponent>
       </approach>
    </extension>
