+++ ccsljava_xdsml/plugins/org.eclipse.gemoc.execution.moccmlxdsml.metaprog/pom.xml
@@ -8,7 +8,7 @@
     <version>2.3.0-SNAPSHOT</version>
   </parent>
 
-  <artifactId>org.eclipse.gemoc.addon.metaprogramming</artifactId>
+  <artifactId>org.eclipse.gemoc.execution.moccmlxdsml.metaprog</artifactId>
   <packaging>eclipse-plugin</packaging>
   <version>2.3.0-SNAPSHOT</version>
 </project>
