+++ ccsljava_xdsml/plugins/org.eclipse.gemoc.execution.moccmlxdsml.metaprog/src/org/eclipse/gemoc/execution/moccmlxdsml/metaprog/MOCCMLRule.java
@@ -1,4 +1,4 @@
-package org.eclipse.gemoc.addon.metaprogramming;
+package org.eclipse.gemoc.execution.moccmlxdsml.metaprog;
 
 import java.util.List;
