+++ releng/org.eclipse.gemoc.execution.concurrent.ccsljavaxdsml.feature/feature.xml
@@ -48,5 +48,11 @@ http://www.eclipse.org/legal/epl-v10.html
          download-size="0"
          install-size="0"
          version="0.0.0"/>
+         
+   <plugin
+         id="org.eclipse.gemoc.execution.moccmlxdsml.metaprog"
+         download-size="0"
+         install-size="0"
+         version="0.0.0"/>
 
 </feature>
